+++ tests/builder.rs
@@ -44,7 +44,7 @@ fn bundle_chain() {
 
         let mut builder = Builder::new(Flags::from_parts(false, true), anchor);
         assert_eq!(
-            builder.add_recipient(None, recipient, NoteValue::from_raw(5000), None),
+            builder.add_output(None, recipient, NoteValue::from_raw(5000), None),
             Ok(())
         );
         let unauthorized = builder.build(&mut rng).unwrap();
@@ -86,7 +86,7 @@ fn bundle_chain() {
         let mut builder = Builder::new(Flags::from_parts(true, true), anchor);
         assert_eq!(builder.add_spend(fvk, note, merkle_path), Ok(()));
         assert_eq!(
-            builder.add_recipient(None, recipient, NoteValue::from_raw(5000), None),
+            builder.add_output(None, recipient, NoteValue::from_raw(5000), None),
             Ok(())
         );
         let unauthorized = builder.build(&mut rng).unwrap();